Move tests from FailureRedeployerTest to DeploymentTriggerTest
FailureRedeployer no longer exists.
